✨ (Revit 2025) - Add Revit 2025 API stubs to available stubs + Add a new shortcut for updating installed stubs
