+++ l10n/bundle.l10n.fr.json
@@ -35,10 +35,12 @@
 
   "🛠️ Select a setup option": "🛠️ Sélectionner une option de configuration",
   "🧩 Install Python extension": "🧩 Installer l'extension Python",
-  "📕 Install Revit API Stubs": "📕 Installer les Stubs Revit API",
+  "📕 Install Revit API Stubs": "📕 Installer les Revit API Stubs",
   "📕 Install pyRevit": "📕 Installer pyRevit",
   "🐍 Install Python": "🐍 Installer Python",
+  "✨ Update Revit API Stubs": "✨ Mettre à jour les Revit API Stubs",
   "📝 Update \"settings.json\"": "📝 Mettre à jour \"settings.json\"",
+
 
   "🎨 Select an icon option": "🎨 Sélectionner une option d'icône",
   "🎨 Enter a color": "🎨 Saisissez une couleur",
