+++ src/lib/commands/setup.ts
@@ -5,6 +5,7 @@ import checkPythonExtension from "../utils/setup/checkPythonExtension";
 import checkPythonInstallation from "../utils/setup/checkPythonInstallation";
 import checkRevitApiStubs from "../utils/setup/checkRevitApiStubs";
 import updateVscSettings from "../utils/setup/updateVscSettings";
+import { updateRevitApiStubs } from "../utils/setup/updateRevitApiStubs";
 
 const { t } = vscode.l10n;
 
@@ -18,6 +19,7 @@ const setup = vscode.commands.registerCommand(
         t("🧩 Install Python extension"),
         t("📕 Install Revit API Stubs"),
         t("📕 Install pyRevit"),
+        t("✨ Update Revit API Stubs"),
         t('📝 Update "settings.json"'),
       ];
 
@@ -38,6 +40,10 @@ const setup = vscode.commands.registerCommand(
             break;
           case t("📕 Install pyRevit"):
             checkPyRevitLib();
+            break;
+          case t("✨ Update Revit API Stubs"):
+            updateRevitApiStubs();
+            updateVscSettings();
             break;
           case t('📝 Update "settings.json"'):
             updateVscSettings();
