+++ src/lib/utils/setup/checkRevitApiStubs.ts
@@ -3,23 +3,13 @@ import * as path from "path";
 import * as vscode from "vscode";
 import revitApiStubsPath from "../../constants/revitApiStubsPath";
 import { showInformationMessage } from "../showMessage";
+import { updateRevitApiStubs } from "./updateRevitApiStubs";
+import updateVscSettings from "./updateVscSettings";
 
 const { t } = vscode.l10n;
 
 const checkRevitApiStubs = () => {
   // Check if Revit API Stubs are installed
-  const sourcePath = path.join(
-    __dirname,
-    "..",
-    "..",
-    "..",
-    "..",
-    "src",
-    "lib",
-    "assets",
-    "RevitAPI stubs"
-  );
-
   if (!fs.existsSync(revitApiStubsPath)) {
     vscode.window
       .showErrorMessage(
@@ -30,11 +20,8 @@ const checkRevitApiStubs = () => {
       )
       .then((action) => {
         if (action === t("Download")) {
-          fs.mkdirSync(revitApiStubsPath, { recursive: true });
-          fs.cpSync(sourcePath, revitApiStubsPath, { recursive: true });
-          showInformationMessage(
-            "The Revit API Stubs are download and place in %APPDATA%."
-          );
+          updateRevitApiStubs();
+          updateVscSettings();
         }
       });
   } else {
